rename + reorganize sliding drawer stuff
