+++ lib/src/main/java/org/oneui/compose/layout/internal/SlidingDrawerLayout.kt
@@ -1,14 +1,9 @@
 package org.oneui.compose.layout.internal
 
-import androidx.compose.animation.core.tween
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
-import androidx.compose.foundation.gestures.AnchoredDraggableState
-import androidx.compose.foundation.gestures.DraggableAnchors
 import androidx.compose.foundation.gestures.Orientation
 import androidx.compose.foundation.gestures.anchoredDraggable
-import androidx.compose.foundation.gestures.animateTo
-import androidx.compose.foundation.gestures.snapTo
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.offset
@@ -34,130 +29,25 @@ enum class SlidingDrawerOpenedState {
 }
 
 /**
- * The state to control the sliding Drawer
- */
-@OptIn(ExperimentalFoundationApi::class)
-data class SlidingDrawerState(
-    val animDuration: Int = 600,
-    val initial: SlidingDrawerOpenedState,
-    val velocityThreshold: Float
-) {
-
-    /**
-     * State for controling the swipeable modifier
-     */
-    internal val draggableState = AnchoredDraggableState(
-        initialValue = initial,
-        positionalThreshold = { distance: Float ->
-            distance / 2F
-        },
-        velocityThreshold = {
-            velocityThreshold
-        },
-        animationSpec = tween(
-            durationMillis = animDuration
-        )
-    )
-
-    fun setAnchors(
-        dragAnchors: Map<SlidingDrawerOpenedState, Float>
-    ) {
-        draggableState.updateAnchors(
-            DraggableAnchors {
-                dragAnchors.forEach {
-                    it.key at it.value
-                }
-            }
-        )
-    }
-
-    /**
-     * Whether the drawer is opened, or whether it was opened before the animation started
-     */
-    val isOpened: Boolean
-        get() = draggableState.currentValue == SlidingDrawerOpenedState.OPEN
-
-    /**
-     * Whether the drawer is closed, or whether it was closed before the animation started
-     */
-    val isClosed: Boolean
-        get() = draggableState.currentValue == SlidingDrawerOpenedState.CLOSED
-
-    /**
-     * Whether the drawer is being swiped at the moment
-     */
-    val isAnimating: Boolean
-        get() = draggableState.isAnimationRunning
-
-    /**
-     * Open drawer with animation
-     */
-    suspend fun openAnimate() = animate(SlidingDrawerOpenedState.OPEN)
-
-    /**
-     * Close drawer with animation
-     */
-    suspend fun closeAnimate() = animate(SlidingDrawerOpenedState.CLOSED)
-
-    /**
-     * Open drawer without animation
-     */
-    suspend fun open() = snap(SlidingDrawerOpenedState.OPEN)
-
-    /**
-     * Close drawer without animation
-     */
-    suspend fun close() = snap(SlidingDrawerOpenedState.CLOSED)
-
-    /**
-     * The progress in pixels the Drawer has made
-     */
-    val pixelProgress: Float
-        get() = draggableState.offset
-
-    private suspend fun animate(
-        target: SlidingDrawerOpenedState
-    ) = draggableState
-        .animateTo(target)
-
-    private suspend fun snap(
-        target: SlidingDrawerOpenedState
-    ) = draggableState
-        .snapTo(target)
-
-}
-
-@Composable
-fun rememberSlidingDrawerState(
-    initial: SlidingDrawerOpenedState = SlidingDrawerOpenedState.CLOSED,
-    velocityThreshold: Float = with(LocalDensity.current) { 100.dp.toPx() }
-) = remember {
-    SlidingDrawerState(
-        initial = initial,
-        velocityThreshold = velocityThreshold
-    )
-}
-
-/**
  * Composable for a Drawerlayout that doesnt overlay the main content but slides it out of the screen.
  *
  * @param modifier The [Modifier] to be applied to the container
  * @param drawerContent The content to be shown inside the Drawer
  * @param contentOverlay The overlay for the main content, depending on the progress. E.g. a scrim.
- * @param state The [SlidingDrawerState] to control the drawer
+ * @param state The [DrawerState] to control the drawer
  * @param maxWidth The width the drawer has when opened
  * @param minWidth The width the drawer has when it is closed
  * @param content The content to be put besides the drawer
  */
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-fun SlidingOutDrawerLayout(
+fun SlidingDrawerLayout(
     modifier: Modifier = Modifier,
     drawerContent: @Composable () -> Unit,
     contentOverlay: @Composable (Float) -> Unit,
     velocityThreshold: Float = with(LocalDensity.current) { 100.dp.toPx() },
-    state : SlidingDrawerState = remember {
-        SlidingDrawerState(
+    state : DrawerState = remember {
+        DrawerState(
             initial = SlidingDrawerOpenedState.CLOSED,
             velocityThreshold = velocityThreshold
         )
@@ -221,8 +111,8 @@ fun SlidingOutDrawerLayout(
 
 @Preview
 @Composable
-fun SidedDrawerLayoutPreview() {
-    SlidingOutDrawerLayout(
+fun SlidingDrawerLayoutPreview() {
+    SlidingDrawerLayout(
         drawerContent = {
             Box(modifier = Modifier.fillMaxSize().background(Color.Blue))
         },
